refactor extern static handling
