+++ src/shims/mod.rs
@@ -10,6 +10,7 @@ pub mod windows;
 mod x86;
 
 pub mod env;
+pub mod extern_static;
 pub mod os_str;
 pub mod panic;
 pub mod time;
